Song huong

+++ EVCare_BackEnd/Application/Interfaces/IConversationService.cs
@@ -12,7 +12,7 @@ namespace Application.Interfaces
         Task<Conversation> CreateOrGetConsultationAsync(string customerAccountId, string staffAccountId);
         Task<string> GetCounterpartAsync(string conversationId, string accountId);
         Task<Dictionary<int, int>> GetUnreadSummaryAsync(string accountId);
-        Task<(List<Conversation>, int, int)> ListMineAsync(string accountId, int pageSize, int pageIndex);
+        Task<(List<Conversation>, int, long)> ListMineAsync(string accountId, int pageSize, int pageIndex);
         Task ResetUnreadAsync(string conversationId, string accountId);
         Task<Conversation> StartConsultationAsync(string customerAccountId, int appointmentId);
     }
